Add files via upload
Lab session #1
II. Getting started with Matlab

II.A. Squares
The goal is to create an image containing centered squares with different gray levels and with one picture as depicted in Figure 1: Squares. The image consists in 4 squares with the following properties:
· First square: dimension 1024×1024, graylevel 32.
· Second square: dimension 512×512, graylevel 64.
· Third square: dimension 256×256, graylevel 128.
· Last square: dimension 128×128, picture 'lena.tif'.

II.B. Mosaic
The goal is to create a mosaic with the different color channels of the Lenna image as depicted in Figure 2: Mosaic with color channels. The different sub-images must be arranged in a single big image, divided as follows:
· Upper left corner: original image (visualized in colour).
· Upper right corner: blue channel after a vertical flip (visualized on blue channel).
· Lower left corner: red channel after a horizontal flip (visualized on red channel).
· Lower right corner: green channel after both flips (visualized on green channel).

II.C. 3D image
Here, you have to display an image as a 3D surface. First, resize the Lenna image half its size and compute its intensity (i.e. convert RGB image to grayscale image). Then plot the image so that the higher the luminance is, the higher the surface is. Furthermore, the plot should display grayscale levels so that the original image can be identified when the surface is observed from above.

II.D. Bitplane slicing
The goal is to create a mosaic containing the bitplanes of the pixel values of Lenna image. Each pixel of the original image is associated with a luminance value which is between 0 and 255. In other terms, each value can be coded using 8 bits. Thus it is possible to create 8 bitplanes gathering all the bit values for a given bit position.